+++ src/skill-manager.ts
@@ -1,12 +1,14 @@
-import {
-    CharacterPF2e,
-    LorePF2e,
-    SkillSlug,
-    ZeroToFour,
-} from "@7h3laughingman/pf2e-types";
+import { CharacterPF2e } from "@7h3laughingman/pf2e-types";
 import { MODULE_ID } from "./module";
-import { migrateData, newData, OneToTwenty, SkillManagerData } from "./data";
-import { isSkill, mapSome, rangeInclusive } from "./utils";
+import {
+    LoreId,
+    migrateData,
+    newData,
+    OneToTwenty,
+    SkillManagerData,
+} from "./data";
+import { mapSome, objectEntries, objectKeys } from "./utils";
+import { LoreInstance, SkillInstance } from "./skill";
 
 export class SkillManager {
     actor: CharacterPF2e;
@@ -43,41 +45,13 @@ export class SkillManager {
     }
 
     prepareData() {
-        const data = this.getData();
-        const ranks =
-            data.increases
-                ?.filter((e) => this.actor.level >= e.level)
-                .reduce(
-                    (acc, e) => {
-                        const upgrade = acc[e.slug];
-                        acc[e.slug] = upgrade
-                            ? (Math.max(upgrade, e.rank) as ZeroToFour)
-                            : e.rank;
-                        return acc;
-                    },
-                    {} as Record<
-                        NonNullable<SkillManagerData["increases"]>[0]["slug"],
-                        ZeroToFour
-                    >,
-                ) ??
-            ({} as Record<
-                NonNullable<SkillManagerData["increases"]>[0]["slug"],
-                ZeroToFour
-            >);
-
         objectKeys(CONFIG.PF2E.skills).forEach((s) => {
-            this.actor.system.skills[s].rank = updateValue(
-                this.actor.system.skills[s].rank,
-                ranks[s],
-                data.overrides?.[s],
-            );
+            const skill = new SkillInstance(s, this);
+            this.actor.system.skills[s].rank = skill.getAppliedRank();
         });
         this.actor.itemTypes.lore.forEach((lore) => {
-            lore.system.proficient.value = updateValue(
-                lore.system.proficient.value,
-                ranks[lore.id],
-                data.overrides?.[lore.id],
-            );
+            const loreInstance = new LoreInstance(lore.id, this);
+            lore.system.proficient.value = loreInstance.getAppliedRank();
         });
     }
 
@@ -112,72 +86,6 @@ export class SkillManager {
         );
     }
 
-    getCumulativeRanks(
-        slug: SkillSlug | LorePF2e["id"],
-    ): Record<
-        | 0
-        | OneToTwenty
-        | "source"
-        | "class"
-        | "background"
-        | "override"
-        | "final",
-        ZeroToFour
-    > {
-        const data = this.getData();
-        const isLore = !isSkill(slug);
-        let source, characterClass, background, final;
-        if (isLore) {
-            const lore = this.actor.itemTypes.lore.find(
-                (lore) => lore.id === slug,
-            );
-            source = lore?._source.system.proficient.value ?? 0;
-            characterClass = 0;
-            background = 0;
-            final = lore?.system.proficient.value ?? 0;
-        } else {
-            source = this.actor._source.system.skills[slug]?.rank ?? 0;
-            characterClass = Number(
-                this.actor.class?.system.trainedSkills.value.includes(slug),
-            );
-            background = Number(
-                this.actor.background?.system.trainedSkills.value.includes(
-                    slug,
-                ),
-            );
-            final = this.actor.system.skills[slug]?.rank ?? 0;
-        }
-        const atZero = Math.max(
-            source,
-            data.settings?.["mark-background-class"] ? background : 0,
-            data.settings?.["mark-background-class"] ? characterClass : 0,
-        ) as ZeroToFour;
-        const increases = this.getData().increases ?? [];
-        const levels = rangeInclusive(1, 20).reduce(
-            (arr, level) => {
-                const inc = increases
-                    .filter((inc) => inc.slug === slug && inc.level === level)
-                    .reduce((rank, e) => Math.max(rank, e.rank), 0);
-                const prevRank = arr[arr.length - 1][1];
-                const rank = Math.max(prevRank, inc) as ZeroToFour;
-                arr.push([level, rank] as [OneToTwenty, ZeroToFour]);
-                return arr;
-            },
-            [[0, atZero]] as [0 | OneToTwenty, ZeroToFour][],
-        );
-        const override =
-            this.getData().overrides?.[slug] ?? levels[levels.length - 1][1];
-
-        return {
-            ...fromEntries(levels),
-            source,
-            class: characterClass as ZeroToFour,
-            background: background as ZeroToFour,
-            override,
-            final,
-        };
-    }
-
     getSkills() {
         return objectEntries(CONFIG.PF2E.skills).map(([slug, { label }]) => ({
             slug: slug,
@@ -186,37 +94,8 @@ export class SkillManager {
     }
     getLores() {
         return this.actor.itemTypes.lore.map((lore) => ({
-            slug: lore.id as LorePF2e["id"],
+            slug: lore.id as LoreId,
             label: lore.name,
         }));
     }
 }
-
-function updateValue(
-    original: ZeroToFour,
-    upgrade?: ZeroToFour,
-    override?: ZeroToFour,
-) {
-    if (typeof override !== "undefined") return override;
-
-    return Math.max(original, upgrade ?? 0) as ZeroToFour;
-}
-
-const objectKeys = <T extends object>(obj: T): (keyof T)[] => {
-    return Object.keys(obj) as (keyof T)[];
-};
-export const objectEntries = <T extends object>(
-    obj: T,
-): [keyof T, T[keyof T]][] => {
-    return Object.entries(obj) as [keyof T, T[keyof T]][];
-};
-
-export const fromEntries = <K extends string | number | symbol, V>(
-    entries: [K, V][],
-): Record<K, V> => {
-    return Object.fromEntries(entries) as Record<K, V>;
-};
-
-export const maxRank = (level: OneToTwenty) => {
-    return level >= 15 ? 4 : level >= 7 ? 3 : level >= 2 ? 2 : 1;
-};